Revert "Release 108"

+++ drivers/hmis/app/graphql/mutations/delete_project_config.rb
@@ -13,7 +13,7 @@ module Mutations
 
     def resolve(id:)
       record = Hmis::ProjectConfig.find(id)
-      access_denied! unless current_user.can_configure_data_collection?
+      raise 'Access denied' unless allowed?(permissions: [:can_configure_data_collection])
 
       record.destroy!
